+++ tests/rules/mod.rs
@@ -5,6 +5,7 @@
 pub mod header_rule_tests;
 pub mod module_registry_rule_tests;
 pub mod readable_source_rule_tests;
+pub mod single_implemented_type_rule_tests;
 pub mod test_file_structure_rule_tests;
 pub mod test_free_source_rule_tests;
 pub mod tests_layout_rule_tests;